+++ src/Components/Editprofile/PostComments.css
@@ -27,6 +27,6 @@
   display: flex;
   gap: 10px;
   align-items: center;
-  width: 850px;
+  width: 800px;
   margin-bottom: 20px;
   }
